some file updated

+++ com/justimagine/Assignement12/Bank.java
@@ -3,10 +3,10 @@ package com.justimagine.Assignement12;
 public abstract class Bank {
 
 	
-double principleAmount=5000;
+float principleAmount=5000;
 	
 	int tenure=5;
-	//float rateOfInterest=5.0f;
 	
-	public abstract void getCalculateInterest(double principleAmount, int tenure) ;
+	
+	public abstract void getCalculateInterest(float principleAmount, int tenure) ;
 }
